% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{Event-class}
\alias{Event-class}
\alias{Event}
\title{Event Class
Create an event emitter and listener}
\description{
Event Class
Create an event emitter and listener
}
\section{Fields}{

\describe{
\item{\code{listeners}}{Registered listeners.}
}}


